@@ -1,12 +1,9 @@
 ﻿using MediatR;
-using SingleApi;
-using System.IO;
+using MetaFile;
 
 namespace Example;
 
-public class FileUpload : IRequest<string>, ISapiFile
+public class FileUpload : StreamFile, IRequest<string>
 {
-    public Stream Content { get; set; } = Stream.Null;
-    public string? Type { get; set; }
-    public string? Name { get; set; }
+
 }
