@@ -1,12 +1,12 @@
 ﻿using MediatR;
-using SingleApi;
+using MetaFile;
 namespace Example.MediatR;
 
-public class FileDownloadHandler : IRequestHandler<FileDownload, SapiFile>
+public class FileDownloadHandler : IRequestHandler<FileDownload, HttpFile>
 {
-    public Task<SapiFile> Handle(FileDownload request, CancellationToken cancellationToken)
+    public Task<HttpFile> Handle(FileDownload request, CancellationToken cancellationToken)
     {
-        return Task.FromResult(new SapiFile()
+        return Task.FromResult(new HttpFile()
         {
             Content = File.OpenRead(Path.GetFullPath(request.Path!)),
             Name = Path.GetFileName(request.Path),
